requirements.txt
slackclient==2.9.4
slack-bolt==1.15.2
python-dotenv==0.21.0
